[#42234] Use only one preopened descriptor
So far, we have been using three preopened descriptors - two for `/` and one for `$HOME`.
We did that because coreutils would break for some reason if we did otherwise.
This commit only leaves a descriptor for `/` and everything works.
If anything breaks in the future, this might be the commit to revert.